refactor(web): collapse [slug]/[locale] + [slug]/[episode]/[locale] into [slug]/[...rest]
Next.js refuses to co-locate sibling dynamic segments with different names
("You cannot use different slug names for the same dynamic path
('episode' !== 'locale')"). Phase 2's two parallel handlers tripped that
constraint at dev-server startup — caught by local smoke test, not by
unit tests which loaded each route in isolation.

Catch-all route replaces both:

  app/[slug]/[locale]/page.tsx        (2-seg, video / series / experience)
  app/[slug]/[episode]/[locale]/page.tsx (3-seg, series-episode)
       →
  app/[slug]/[...rest]/page.tsx       (segment-count dispatcher)

Segment-count branches:
  - rest.length === 1 → 2-seg behavior (rest[0] = locale)
  - rest.length === 2 → 3-seg behavior (rest[0] = episode, rest[1] = locale)
  - other → notFound()

Two render helpers (renderVideo, renderEpisode) inline the prior handlers
verbatim — same Experience precedence, same resync redirect (.html shape
via watchVideoPath / watchEpisodePath builders), same SeriesPageClient /
WatchPageClient dispatch. generateMetadata gets the same dispatch.

Adds future-proofing: a new URL shape (4-seg) is one more branch + one
more ParsedWatchPath kind. No more "Next.js refuses to start" surprise.

Also: moved typedRoutes from experimental to top-level (stable in Next 16
per the dev-server warning).

Tests: merged into a single page-routing.test.tsx with describe-blocks
per kind. 22 tests (up from 14 + 7 across the prior two files). Full
apps/web suite: 56 files, 786 tests + 2 todos. Typecheck + lint clean.

Smoke QA: started web on :3010 against a local admin GraphQL endpoint;
verified all Phase 2 URL shapes resolve 200 with real content:
  /watch                                                  → home
  /watch/videos                                           → All Videos
  /watch/jesus.html                                       → Jesus
  /watch/jesus.html/english.html                          → Jesus
  /watch/jesus.html/spanish-castilian.html                → Jesus
  /watch/lumo-the-gospel-of-john.html/wedding-in-cana/english.html → Wedding in Cana
  /watch/jesus (bare transitional)                        → Jesus
  /watch/a/b/c/d (4-seg unknown)                          → notFound page
