+++ apps/web/src/app/[slug]/[...rest]/error.tsx
