+++ apps/web/src/app/[slug]/[...rest]/loading.tsx
