--- apps/web/src/app/[slug]/[episode]/[locale]/loading.tsx
@@ -1,5 +1,0 @@
-import { ExperienceSkeleton } from "@/components/ExperienceSkeleton"
-
-export default function Loading() {
-  return <ExperienceSkeleton />
-}
